refactor(web): merge micro theme overrides; single-source palette hexes
