+++ apps/web/src/app/manifest.ts
@@ -1,4 +1,5 @@
 import type { MetadataRoute } from "next";
+import { surfaces } from "@/theme/palette";
 
 export default function manifest(): MetadataRoute.Manifest {
   return {
@@ -8,8 +9,8 @@ export default function manifest(): MetadataRoute.Manifest {
       "Drive Claude Code or Codex on your own subscription to search job boards, tailor your resume, apply, and track every reply.",
     start_url: "/",
     display: "standalone",
-    background_color: "#0B0B0A",
-    theme_color: "#0B0B0A",
+    background_color: surfaces.base,
+    theme_color: surfaces.base,
     icons: [{ src: "/icon.svg", type: "image/svg+xml", sizes: "any", purpose: "any" }],
   };
 }
