--- apps/web/src/theme/overrides/paper.ts
@@ -1,11 +1,0 @@
-import type { Components, Theme } from "@mui/material/styles";
-
-export const paperOverrides: Components<Theme>["MuiPaper"] = {
-  defaultProps: { elevation: 0 },
-  styleOverrides: {
-    root: ({ theme }) => ({
-      backgroundImage: "none",
-      backgroundColor: theme.palette.surfaces.card,
-    }),
-  },
-};
