--- apps/web/src/theme/overrides/typography.ts
@@ -1,13 +1,0 @@
-import type { Components, Theme } from "@mui/material/styles";
-
-export const typographyOverrides: Components<Theme>["MuiTypography"] = {
-  defaultProps: {
-    variantMapping: {
-      body1Muted: "p",
-      body2Muted: "p",
-      captionMuted: "span",
-      overline: "span",
-      overlineMuted: "span",
-    },
-  },
-};
