+++ apps/web/src/theme/palette.ts
@@ -43,11 +43,11 @@ export const stages = {
 } as const;
 
 export const editorial = {
-  paper: "#F4F2EE",
-  ink: "#0B0B0A",
-  thrust: "#3B82F6",
-  flame: "#FF6A3D",
-  amber: "#FFB020",
+  paper: textColors.primary,
+  ink: surfaces.base,
+  thrust: accent.secondary,
+  flame: accent.primary,
+  amber: feedback.warning,
 } as const;
 
 export type StageKey = keyof typeof stages;
